Generate library from shelley.cddl + hash structs
This uses just a subset of shelley.cddl to generate that we'll need to
making transactions so we can calculate fees later.

As we interpret everything as just bytes for hashes, instead we'll just
write wrappers that are exact-sized and do checks for that using a
macro.

AddrCred/AddrCredType were also basically the same thing as the
`stake_credential` in shelley.cddl so those were merged into one as
well and the outside type renamed.

js_chain_libs.rs -> crypto.rs to keep lib.rs/serialization.rs/prelude.rs
as just the generated code from the CDDL.