Add Coin wrapper + Refactor MoveInstantaneousRewards  (#14)
* Use from_bytes() instead of from_bytes_impl() in tests

This is the user-facing interface so it makes more sense to test I
think.

* Give MoveInstantaneousReward proper names

In the last shelley.cddl re-gen I didn't notice
move_instantaneous_reward added in a source enum but it's just specified
as `0 / 1` in the spec which gives really ugly confusing code so I just
made it a proper enum and cleaned up the classes and merged the two.

* Make Coin its own class

Instead of just a u64, which is converted to BigNum in wasm, we have our
own wrapper type similar to js-chain-libs.

The reason for this is that BigNUm might not exist in all frameworks
(e.g. ReactNative) so we need to have the ability to create/get them as
strings.