AuxilaryData implementation
Like the Shelley -> Shelley Mary with native scripts, the type
definition is quite awkward with multiple types when it boils down to
simply having optional fields, so instead of working with super ugly
codegened types with like 5 types instead of 1, I just hand-wrote this
code.

I kept the old shelley / mary formats for serialization as they reduce
on-chain space usage, and they're required to be implemented for
deserialization anyway in alonzo.

I didn't keep the old TransacitonMetadata name as I think it is
different enough that it makes sense to break older code which would
likely have to be updated anyway to go through with the AuxilaryData
name IOHK is using in their binary spec now, rather than storing plutus
stuff inside of the "metadata".